pyproject.toml
[tool.poetry]
name = "aia-agent"
version = "0.1.0"
description = "Agentes inteligentes para automatización"
authors = ["Edgar"]
readme = "README.md"
packages = [{include = "workers"}]

[tool.poetry.dependencies]
python = "^3.11"
dash = "^2.9.3"
plotly = "^5.15.0"
paho-mqtt = "^1.6.1"
pymongo = "^4.3.3"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
mqtt-worker = "workers.mqtt_worker:main"
whatsapp-reader = "workers.whatsapp_reader:main"
whatsapp-sender = "workers.whatsapp_sender:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
